Double quality depreciation after sell-by date.
There's an argument to be made that I should have calculated the
depreciation factor in-line and then factored it out as a method later.
I did it this way because it mirrored my thinking process: "we need to
multiply this number by a factor here, let's work out the factor
somewhere else". Not "we need to do some extra logic here".

Notice that `starting_sell_by` now defaults to Float::INFINITY instead
of zero, in order to keep older tests passing. This would have been a
better default to start with anyway. It's more "benign" in this context
than 0 is.