Handle edge case for max aged brie quality.
Notice the adjustment down to the smaller of current or max
value, *after* value is potentially incremented beyond the max.

This is stinky from a constraints-oriented, immutable, functional
perspective. But I sometimes find it the easiest way to reason about a
problem. It neatly separates responsibilities from line to line: the
first line is unconcerned with max, and the second line is unconcerned
with increment rules.

Think of it as treating the middle of a method, and the variables it
deals with, as a "scratchpad". They can have "wrong" intermediate
values, so long as they end up in the right place.